Move SamInitRunner to the project wizard instead of lambda.executation package
